split code to new file, test=asr
